+++ lib/templates/deploy/render/render-build.sh
@@ -1,0 +1,10 @@
+#!/usr/bin/env bash
+set -euo pipefail
+
+echo "--> Checking environment"
+echo "RODBOT_ENV: $RODBOT_ENV"
+echo "RODBOT_PLUGINS: $RODBOT_PLUGINS"
+
+echo "--> Installing bundle"
+bundle config set with $RODBOT_PLUGINS
+bundle install