+++ lib/templates/deploy/render/render-start.sh
@@ -1,0 +1,5 @@
+#!/usr/bin/env bash
+
+export RODBOT_APP_HOST=0.0.0.0
+bundle exec rodbot start
+sleep infinity